fixed code based on comments from the pull request review
