--- RIckYMortyUIKIt....xcodeproj/xcuserdata/paulafernandezvazquez.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "1EBDC905-997D-44A5-9A9B-014F7052A0B3"
-   type = "1"
-   version = "2.0">
-</Bucket>
