--- RIckYMortyUIKIt.../.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  RIckYMortyUIKIt...
-//
-//  Created by Paula Fernández Vázquez on 23/5/25.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
