--- RIckYMortyUIKIt.../Model/LocationModel.swift
@@ -1,13 +1,0 @@
-//
-//  mod.swift
-//  ProyectoRick
-//
-//  Created by Paula Fernández Vázquez on 16/5/25.
-//
-
-import Foundation
-
-struct Location: Codable {
-    let name: String
-    let url: String
-}
