--- RIckYMortyUIKIt.../Model/OriginModel.swift
@@ -1,13 +1,0 @@
-//
-//  OrigenModel.swift
-//  ProyectoRick
-//
-//  Created by Paula Fernández Vázquez on 16/5/25.
-//
-
-import Foundation
-
-struct Origin: Codable {
-    let name: String
-    let url: String
-}
