+++ Ricky-Morty-PLX/Assets.xcassets/Contents.json
