+++ Ricky-Morty-PLX/Info.plist
