+++ Ricky-Morty-PLX/Model/APICommunication/ApiServiceProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  APIProtocol.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Ainhoa  on 26/5/25.
+//
+
+import Foundation
+import Combine
+
+protocol ApiServiceProtocol {
+    func recieveCharacters(url: String?) -> AnyPublisher<CharacterResultModel, Error>
+}
