+++ Ricky-Morty-PLX/Model/CharacterModel.swift
@@ -1,0 +1,19 @@
+//
+//  CharacterModel.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Ainhoa  on 26/5/25.
+//
+
+import Foundation
+
+struct Character: Codable,Identifiable {
+    let id: Int
+    let name: String
+    let status: String
+    let species: String
+    let gender: String
+    let origin: Origin
+    let location: Location
+    let image: String
+}
