+++ Ricky-Morty-PLX/Model/CharacterResultModel.swift
@@ -1,0 +1,13 @@
+//
+//  CharacterResult.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Ainhoa  on 26/5/25.
+//
+
+import Foundation
+
+struct CharacterResultModel: Codable {
+    let info: InfoModel
+    let results: [Character]
+}
