+++ Ricky-Morty-PLX/Model/InfoModel.swift
@@ -1,0 +1,12 @@
+//
+//  InfoModel.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Paula Fernández Vázquez on 27/5/25.
+//
+
+import Foundation
+
+struct InfoModel: Codable {
+    let next: String?
+}
