+++ Ricky-Morty-PLX/Model/LocationModel.swift
@@ -1,0 +1,13 @@
+//
+//  LocationModel.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Ainhoa  on 26/5/25.
+//
+
+import Foundation
+
+struct Location: Codable {
+    let name: String
+    let url: String
+}
