+++ Ricky-Morty-PLX/Model/OriginModel.swift
@@ -1,0 +1,13 @@
+//
+//  OriginModel.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Ainhoa  on 26/5/25.
+//
+
+import Foundation
+
+struct Origin: Codable {
+    let name: String
+    let url: String
+}
