+++ Ricky-Morty-PLX/Model/PaginationModel.swift
@@ -1,0 +1,13 @@
+//
+//  PaginationModel.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Paula Fernández Vázquez on 27/5/25.
+//
+
+import Foundation
+
+struct PaginationModel: Codable {
+    let next: String?
+    let prev: String?
+}
