+++ Ricky-Morty-PLX/View/CharactersCellProtocol.swift
@@ -1,0 +1,12 @@
+//
+//  CharactersCellProtocol.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Paula Fernández Vázquez on 27/5/25.
+//
+
+import Foundation
+
+protocol CharactersCellProtocol {
+    func configure(character: Character)
+}
