+++ Ricky-Morty-PLX/View/CharatersListView.swift
@@ -1,0 +1,90 @@
+//
+//  CharatersListView.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Ainhoa  on 26/5/25.
+//
+import UIKit
+import Combine
+
+class CharactersListView: UIViewController, UICollectionViewDelegate, UICollectionViewDataSource {
+    private var viewModel: CharactersListViewModel
+    private var collectionView: UICollectionView!
+    private var cancelables = Set<AnyCancellable>()
+    private let configureUseCase = ConfigureUseCase(apiService: ApiCall())
+
+
+    init(viewModel: CharactersListViewModel) {
+        self.viewModel = viewModel
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Rick&Morty Characters"
+        view.backgroundColor = .darkGray
+        configureCollectionView()
+        linkViewModel()
+        viewModel.loadCharacters()
+    }
+
+    private func configureCollectionView() {
+        let layout = UICollectionViewFlowLayout()
+        layout.itemSize = CGSize(width: 150, height: 200)
+        layout.sectionInset = UIEdgeInsets(top: 10, left: 10, bottom: 10, right: 10)
+        layout.minimumLineSpacing = 15
+
+        collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: layout)
+        collectionView.backgroundColor = .darkGray
+        collectionView.autoresizingMask = [.flexibleWidth, .flexibleHeight]
+        collectionView.register(CharactersCell.self, forCellWithReuseIdentifier: CharactersCell.reuseId)
+        collectionView.delegate = self
+        collectionView.dataSource = self
+
+        view.addSubview(collectionView)
+    }
+
+    private func linkViewModel() {
+        viewModel.$characters
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] _ in
+                self?.collectionView.reloadData()
+            }
+            .store(in: &cancelables)
+    }
+
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return viewModel.count
+    }
+
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(
+            withReuseIdentifier: CharactersCell.reuseId,
+            for: indexPath
+        ) as? CharactersCell else {
+            return UICollectionViewCell()
+        }
+        let character = viewModel.getCharacters(index: indexPath.item)
+        cell.injectUseCase(useCase: configureUseCase)
+        cell.configure(character: character)
+        return cell
+    }
+    
+    // TODO AÑADIR FUNCIONALIDAD PARA IR A LA CLASE CHARACTERDETAILS
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        let selectedCharacter = viewModel.characters[indexPath.item]
+        let vcDetail = CharacterDetailView(character: selectedCharacter)
+        navigationController?.pushViewController(vcDetail, animated: true)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
+        let lastItem = viewModel.characters.count - 1
+        if indexPath.item == lastItem {
+            viewModel.loadCharacters()
+        }
+    }
+}
