+++ Ricky-Morty-PLX/ViewModel/CharacterService.swift
@@ -1,0 +1,12 @@
+//
+//  CharacterService.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Paula Fernández Vázquez on 27/5/25.
+//
+
+import Foundation
+
+protocol CharacterService {
+    func loadCharacters()
+}
