+++ Ricky-Morty-PLX/ViewModel/CharactersListViewModelProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  ViewModelProtocol.swift
+//  Ricky-Morty-PLX
+//
+//  Created by Ainhoa  on 26/5/25.
+//
+
+import Foundation
+
+protocol charactersListViewModelProtocol {
+    func loadCharacters()
+    func getCharacters(index: Int) -> Character
+}
